+++ app/controllers/search_controller.rb
@@ -10,7 +10,7 @@ class SearchController < ApplicationController
     @member = User.current.find_by("LOWER(username) = ?", params[:search].to_s.downcase)
     @search = clean_search.split(/[^\w]/).reject(&:blank?).uniq.join(" & ")
     @search_documents = PgSearch.multisearch(params[:search]).page(params[:page]).per(10)
-    if clean_search.present?
+    if clean_search.present? && params[:page].blank?
       search = Search.where(search: clean_search).first_or_create
       search.update results_count: @search_documents.total_count
       search.increment! :search_count